Display embedded licence MD file as rendered HTML (#7975)
