+++ src/NuGetGallery/Views/Packages/License.cshtml
@@ -66,7 +66,14 @@
                 {
                     <label>License text</label>
                     <div class="common-licenses">
-                        <pre class="license-file-contents custom-license-container">@Model.LicenseFileContents</pre>
+                        @if (Model.LicenseFileContentsHtml != null)
+                        {
+                            @Html.Raw(Model.LicenseFileContentsHtml)
+                        } 
+                        else 
+                        {
+                            <pre class="license-file-contents custom-license-container">@Model.LicenseFileContents</pre>
+                        }
                     </div>
                 }
             }
