+++ src/NuGetGallery/Views/Packages/_VerifyMetadata.cshtml
@@ -131,12 +131,15 @@
             </div>
             <!-- /ko -->
 
-            <!-- ko if: $data.LicenseFileContents -->
             <div class="verify-package-field common-licenses">
                 <label class="verify-package-field-heading">License file</label>
+                <!-- ko if: $data.LicenseFileContentsHtml -->
+                <span data-bind="html: $data.LicenseFileContentsHtml"></span>
+                <!-- /ko -->
+                <!-- ko ifnot: $data.LicenseFileContentsHtml -->
                 <pre class="license-file-contents custom-license-container" data-bind="text: $data.LicenseFileContents"></pre>
+                <!-- /ko -->
             </div>
-            <!-- /ko -->
 
             <!-- ko if: !$data.LicenseExpression && !$data.LicenseFileContents && !$data.LicenseExpressionSegments -->
             <div class="verify-package-field">
